fix(render): fix formatting float amount

+++ cmd/app/render.go
@@ -28,7 +28,7 @@ var functions = template.FuncMap{
 }
 
 func formatAmount(a int) string {
-	f := float32(a / 100)
+	f := float32(a) / float32(100)
 	return fmt.Sprintf("$%.2f", f)
 }
